requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
deta==1.1.0
fastapi==0.85.1
h11==0.13.0
idna==3.3
iniconfig==1.1.1
mypy==0.982
mypy-extensions==0.4.3
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.3
python-decouple==3.6
requests==2.28.1
sniffio==1.2.0
starlette==0.20.4
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.2.0
urllib3==1.26.12
uvicorn==0.17.6
